version 1.3.0 Updated version to 1.19 moved files to io.github.leguan16.smartinvs
