+++ pom.xml
@@ -6,9 +6,9 @@
 
     <groupId>io.github.leguan16</groupId>
     <artifactId>smartInvs</artifactId>
-    <version>1.2.0</version>
+    <version>1.3.0</version>
 
-    <name>SmartInvs-1.17.1</name>
+    <name>SmartInvs-1.19</name>
 
     <description>A more recent and updated version of SmartInvs by MinusKube</description>
     <url>https://github.com/Leguan16/SmartInvs</url>
@@ -164,7 +164,7 @@
         <dependency>
             <groupId>io.papermc.paper</groupId>
             <artifactId>paper-api</artifactId>
-            <version>1.17.1-R0.1-SNAPSHOT</version>
+            <version>1.19-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
