+++ src/main/java/io/github/leguan16/smartinvs/ClickableItem.java
@@ -1,4 +1,4 @@
-package io.github.leguan16;
+package io.github.leguan16.smartinvs;
 
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
