+++ src/main/java/io/github/leguan16/smartinvs/opener/InventoryOpener.java
@@ -1,8 +1,8 @@
-package io.github.leguan16.opener;
+package io.github.leguan16.smartinvs.opener;
 
-import io.github.leguan16.ClickableItem;
-import io.github.leguan16.SmartInventory;
-import io.github.leguan16.content.InventoryContents;
+import io.github.leguan16.smartinvs.ClickableItem;
+import io.github.leguan16.smartinvs.SmartInventory;
+import io.github.leguan16.smartinvs.content.InventoryContents;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.inventory.Inventory;
