+++ src/main/resources/plugin.yml
@@ -7,4 +7,4 @@ authors:
   - MinusKube
   - MCmitNoah
 website: https://github.com/MinusKube/SmartInvs
-main: io.github.leguan16.SmartInvsPlugin
+main: io.github.leguan16.smartinvs.SmartInvsPlugin
